+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fhevmsdk/core
 
+## 0.2.0
+
+### Minor Changes
+
+- Rename packages name
+
 ## 0.1.0
 
 ### Minor Changes