+++ packages/react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @fhevmsdk/react
 
+## 0.2.0
+
+### Minor Changes
+
+- Rename packages name
+
+### Patch Changes
+
+- Updated dependencies
+  - @fhevmsdk/core@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes